Extract FilenameHelper from Core\Utils (#563 Sprint 3, PR 3a)
Begin dismantling the Core\Utils god-utility (364 inbound refs) by moving
the filename builders into a focused FreeFormCertificate\Core\FilenameHelper:
sanitize_filename(), build_pdf_filename(), get_export_filename(). All call
sites repointed (no back-compat wrappers); pure string helpers, behavior-
preserving. The request-input accessors follow in PR 3b.
