+++ includes/reregistration/class-ffc-reregistration-csv-exporter.php
@@ -64,7 +64,7 @@ class ReregistrationCsvExporter {
 		$fields      = self::get_custom_fields_for_reregistration( $rereg );
 
 		// Build CSV.
-		$filename = \FreeFormCertificate\Core\Utils::get_export_filename( 'reregistration', (string) $rereg->title );
+		$filename = \FreeFormCertificate\Core\FilenameHelper::get_export_filename( 'reregistration', (string) $rereg->title );
 
 		// Headers.
 		$safe_filename = str_replace( array( "\r", "\n", '"' ), '', $filename );
